Play Video Script and render texture
Script that can store all the video clips for the cartridges, to be attached to the console playing it. Needs video player object in the scene to reference too and to be connected to the render texture. Contains a videoClipIndex value that needs to be defined by the specific cartridges and then the function playVideo to be used in the interact script.